@@ -3,7 +3,6 @@ package me.sunzheng.mana
 import android.app.SearchManager
 import android.content.ComponentName
 import android.content.Intent
-import android.content.res.Configuration
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.net.Uri
@@ -20,110 +19,85 @@ import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.appcompat.widget.SearchView
-import androidx.appcompat.widget.Toolbar
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.core.view.GravityCompat
 import androidx.core.view.MenuItemCompat
 import androidx.databinding.DataBindingUtil
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentStatePagerAdapter
 import androidx.recyclerview.widget.RecyclerView
+import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestBuilder
 import com.bumptech.glide.request.RequestOptions
 import com.google.android.material.navigation.NavigationView
-import com.google.android.material.tabs.TabLayout.TabLayoutOnPageChangeListener
+import com.google.android.material.tabs.TabLayoutMediator
+import dagger.hilt.android.AndroidEntryPoint
 import me.sunzheng.mana.core.AnnounceModel
-import me.sunzheng.mana.core.net.v2.ApiService
 import me.sunzheng.mana.databinding.ActivityMainBinding
-import me.sunzheng.mana.home.HomeApiService
-import me.sunzheng.mana.home.HomeContract
 import me.sunzheng.mana.home.onair.OnAirFragment
-import me.sunzheng.mana.home.onair.OnAirPresenterImpl
-import me.sunzheng.mana.home.onair.respository.DataRepositoryImpl
-import me.sunzheng.mana.utils.App
 import me.sunzheng.mana.utils.HostUtil
 import me.sunzheng.mana.utils.PreferenceManager
 import me.sunzheng.mana.utils.RegexUtils
+import javax.inject.Inject
 
-class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
+@AndroidEntryPoint
+class MainActivity @Inject constructor() : AppCompatActivity(),
+    NavigationView.OnNavigationItemSelectedListener {
     companion object {
         const val CLICK_DELAY_MILLIONSECONDS: Long = 500
     }
 
-    var apiService: ApiService? = null
     var titles: Array<CharSequence>? = null
     var handler = Handler(Looper.getMainLooper())
 
 
     var isJaFirst = false
-    val binder: ActivityMainBinding by lazy {
+    val binding: ActivityMainBinding by lazy {
         DataBindingUtil.setContentView(this, R.layout.activity_main)
     }
+    private val fragmentPagerAdapter2: FragmentStateAdapter by lazy {
+        object : FragmentStateAdapter(this) {
+            override fun getItemCount(): Int = 2
 
-
-    var fragmentPagerAdapter: FragmentStatePagerAdapter = object : FragmentStatePagerAdapter(
-        supportFragmentManager
-    ) {
-        override fun finishUpdate(container: ViewGroup) {
-            super.finishUpdate(container)
-        }
-
-        override fun instantiateItem(container: ViewGroup, position: Int): Any {
-            val fragment = super.instantiateItem(container, position) as OnAirFragment
-            val presenter: HomeContract.OnAir.Presenter =
-                OnAirPresenterImpl(fragment, DataRepositoryImpl(this@MainActivity, apiService))
-            fragment.setPresenter(presenter)
-            presenter.load(fragment.requireArguments().getInt(OnAirFragment.INT_ARGS_TYPE))
-            return fragment
-        }
-
-        override fun getItem(position: Int): Fragment {
-            return OnAirFragment.newInstance(if (position % 2 == 0) OnAirFragment.INT_TYPE_ANIMATION else OnAirFragment.INT_TYPE_DRAMA)
-        }
-
-        override fun getCount(): Int {
-            return 2
-        }
-
-        override fun getPageTitle(position: Int): CharSequence? {
-            return titles?.get(position)
+            override fun createFragment(position: Int): Fragment =
+                OnAirFragment.newInstance(if (position == 0) 2 else 6)
         }
     }
-
 
     fun showContentView(visible: Boolean) {
         showCollapsingToolbarLayout(visible)
         if (visible) {
-            binder.mainView.mainAnnounceRecyclerview!!.visibility = View.VISIBLE
+            binding.mainView.mainAnnounceRecyclerview.visibility = View.VISIBLE
         } else {
-            binder.mainView.mainAnnounceRecyclerview!!.visibility = View.GONE
+            binding.mainView.mainAnnounceRecyclerview.visibility = View.GONE
         }
-        binder.mainView.mainAnnounceView.visibility =
-            binder.mainView.mainAnnounceRecyclerview.visibility
+        binding.mainView.mainAnnounceView.visibility =
+            binding.mainView.mainAnnounceRecyclerview.visibility
     }
 
     fun showCollapsingToolbarLayout(active: Boolean) {
-        if (binder.mainView.appbarlayout == null) return
-        binder.mainView.appbarlayout.setExpanded(active, true)
+        if (binding.mainView.appbarlayout == null) return
+        binding.mainView.appbarlayout.setExpanded(active, true)
     }
 
     fun setData(datas: List<AnnounceModel>) {
-        if (binder.mainView.mainAnnounceRecyclerview != null) binder.mainView.mainAnnounceRecyclerview!!.adapter =
+        if (binding.mainView.mainAnnounceRecyclerview != null) binding.mainView.mainAnnounceRecyclerview.adapter =
             AnnouceAdapter(datas)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        val toolbar = findViewById<View>(R.id.toolbar) as Toolbar
-        setSupportActionBar(toolbar)
+        setSupportActionBar(binding.mainView.toolbar)
         supportActionBar!!.setDisplayShowTitleEnabled(true)
         supportActionBar!!.setTitle(R.string.app_name)
         val drawer = findViewById<View>(R.id.drawer_layout) as DrawerLayout
         val toggle: ActionBarDrawerToggle = object : ActionBarDrawerToggle(
-            this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close
+            this,
+            drawer,
+            binding.mainView.toolbar,
+            R.string.navigation_drawer_open,
+            R.string.navigation_drawer_close
         ) {
             override fun onDrawerSlide(drawerView: View, slideOffset: Float) {
 //              see https://stackoverflow.com/questions/27117243/disable-hamburger-to-back-arrow-animation-on-toolbar
@@ -146,36 +120,25 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
                     PreferenceManager.Global.RES_JA_FIRST_BOOL
                 ), false
             )
-        apiService = (application as App).retrofit.create(
-            ApiService::class.java
-        )
         titles = arrayOf(
             getText(R.string.title_anim_catalog_tablayout),
             getText(R.string.title_dram_catalog_tablayout)
         )
-        binder.mainView.mainViewpager.adapter = fragmentPagerAdapter
-        binder.mainView.mainViewpager.addOnPageChangeListener(TabLayoutOnPageChangeListener(binder.mainView.mainTablayout))
-        binder.mainView.mainTablayout.setupWithViewPager(binder.mainView.mainViewpager, true)
-//        setPresenter(
-//            AnnoucePresenterImpl(
-//                this, (applicationContext as App).retrofit.create(
-//                    Announce::class.java
-//                )
-//            )
-//        )
-//        mPresenter!!.load()
+        binding.mainView.mainViewpager.adapter = fragmentPagerAdapter2
+        TabLayoutMediator(
+            binding.mainView.mainTablayout,
+            binding.mainView.mainViewpager
+        ) { tab, position ->
+            tab.text = titles?.get(position) ?: ""
+        }.attach()
     }
 
     override fun onBackPressed() {
-        if (binder.drawerLayout.isDrawerOpen(GravityCompat.START)) {
-            binder.drawerLayout.closeDrawer(GravityCompat.START)
+        if (binding.drawerLayout.isDrawerOpen(GravityCompat.START)) {
+            binding.drawerLayout.closeDrawer(GravityCompat.START)
         } else {
             super.onBackPressed()
         }
-    }
-
-    override fun onConfigurationChanged(newConfig: Configuration) {
-        super.onConfigurationChanged(newConfig)
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -340,5 +303,4 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
             }
         }
     }
-
 }
